added error if data=0

+++ internal/repository/postgres/repository.go
@@ -129,6 +129,14 @@ func (r *Repository) SubscribeToSociety(id int64, uuid string) (bool, error) {
 }
 
 func (r *Repository) UnsubscribeFromSociety(id int64, uuid string) (bool, error) {
+	var data []model.IdSociety
+	err := r.connection.Select(&data, "SELECT id FROM societies_subscribers WHERE id = $1 AND user_uuid = $2", id, uuid)
+	if err != nil {
+		return false, fmt.Errorf("failed to select subscribe from society: %v", err)
+	}
+	if len(data) == 0 {
+		return false, nil
+	}
 	if _, err := r.connection.Exec("DELETE FROM societies_subscribers WHERE society_id = $1 AND user_uuid = $2", id, uuid); err != nil {
 		return false, fmt.Errorf("r.connection.Exec: %v", err)
 	}
